Moved common manifest-related features from the composer and node linter into a manifest mixin
